requirements.txt
fastapi
uvicorn
pydantic
pyjwt
firebase-admin
google-cloud-firestore
